Remove AuditableEntitySaveChangesInterceptor class
This commit removes the `AuditableEntitySaveChangesInterceptor` class and all its associated methods and properties from the `AuditableEntitySaveChangesInterceptor.cs` file. The interceptor, which was responsible for setting audit properties for `IAuditableEntity` instances during `SaveChanges` operations, is no longer present.

Additionally, the registration of the interceptor in the `AddEntityFrameworkDataAccess` method of `ServiceCollectionExtensions.cs` has been commented out, effectively disabling its functionality in the application.